(erlang) Preserve unknown Protobuf fields
Store unknown (unrecognized) Protobuf fields during de-serialization in a
special 'piqi_protobuf_unknown_field_list' field and automatically write them
back when record is serialized to Protobuf.

In addition, unknown Protobuf fields can be easily accessed and manipulated
using standard interfaces from the piqirun module.

Such treatment of unknown fields is consistent with Protocol Buffers behavior.
For instance, its manual states that "... the unknown fields are not discarded,
and if the message is later serialized, the unknown fields are serialized along
with it – so if the message is passed on to new code, the new fields are still
available. Note that preservation of unknown fields is currently not available
for Python" (https://developers.google.com/protocol-buffers/docs/proto). Also,
Protobuf offsers a comprehensive interface for dealing with unknown fields:
https://developers.google.com/protocol-buffers/docs/reference/cpp/google.protobuf.unknown_field_set#UnknownFieldSet